refactor: remove tanstack query
