--- src/hooks/mutations/useRefreshToken.ts
@@ -1,8 +1,0 @@
-import { refreshToken } from '@/api/auth/refreshToken';
-import { useMutation } from '@tanstack/react-query';
-
-export const useRefreshToken = () => {
-  return useMutation({
-    mutationFn: refreshToken,
-  });
-};
